Simplify event manager
