+++ src/utils/eventManager.js
@@ -1,24 +1,24 @@
 const eventManager = {
-  eventList: new Map(),
+  list: new Map(),
 
   on(event, callback) {
-    this.eventList.has(event) || this.eventList.set(event, []);
+    this.list.has(event) || this.list.set(event, []);
 
-    this.eventList.get(event).push(callback);
+    this.list.get(event).push(callback);
 
     return this;
   },
 
-  off(event = null) {
-    return this.eventList.delete(event);
+  clear() {
+    this.list = new Map();
   },
 
   emit(event, ...args) {
-    if (!this.eventList.has(event)) {
+    if (!this.list.has(event)) {
       return false;
     }
 
-    this.eventList
+    this.list
       .get(event)
       .forEach(callback => setTimeout(() => callback.call(this, ...args), 0));
